+++ assets/javascripts/kitten/components/information/alert/test.js
@@ -7,16 +7,22 @@ describe('<Alert />', () => {
     const component = renderer.create(<Alert>Alert !</Alert>).toJSON()
     expect(component).toMatchSnapshot()
   })
-  it('should matches with error snapshot', () => {
-    const component = renderer.create(<Alert error>Alert !</Alert>).toJSON()
+  it('should matches with snapshot with danger status prop', () => {
+    const component = renderer
+      .create(<Alert status="danger">Alert !</Alert>)
+      .toJSON()
     expect(component).toMatchSnapshot()
   })
-  it('should matches with success snapshot', () => {
-    const component = renderer.create(<Alert success>Alert !</Alert>).toJSON()
+  it('should matches with snapshot with success status prop', () => {
+    const component = renderer
+      .create(<Alert status="success">Alert !</Alert>)
+      .toJSON()
     expect(component).toMatchSnapshot()
   })
-  it('should matches with warning snapshot', () => {
-    const component = renderer.create(<Alert warning>Alert !</Alert>).toJSON()
+  it('should matches with snapshot with warning status prop', () => {
+    const component = renderer
+      .create(<Alert status="warning">Alert !</Alert>)
+      .toJSON()
     expect(component).toMatchSnapshot()
   })
   it('should matches with closeButton snapshot', () => {
@@ -26,6 +32,24 @@ describe('<Alert />', () => {
           Alert !
         </Alert>,
       )
+      .toJSON()
+    expect(component).toMatchSnapshot()
+  })
+  it('should matches with snapshot without icon', () => {
+    const component = renderer
+      .create(<Alert displayIcon={false}>Alert !</Alert>)
+      .toJSON()
+    expect(component).toMatchSnapshot()
+  })
+  it('should matches with snapshot with icon border', () => {
+    const component = renderer
+      .create(<Alert iconHasBorder>Alert !</Alert>)
+      .toJSON()
+    expect(component).toMatchSnapshot()
+  })
+  it('should matches with snapshot with custom icon', () => {
+    const component = renderer
+      .create(<Alert icon={<svg />}>Alert !</Alert>)
       .toJSON()
     expect(component).toMatchSnapshot()
   })